+++ lib/assets/javascripts/new-dashboard/components/MapsList.vue
@@ -58,7 +58,10 @@
       <div
           v-if="shouldShowListHeader"
           class="grid-cell grid-cell--noMargin grid-cell--col12 grid__head--sticky"
-          :class="{ 'is-user-notification': isNotificationVisible }">
+          :class="{
+            'has-userNotification': isNotificationVisible,
+            'in-homePage': isHomePage,
+          }">
         <CondensedMapHeader
           :order="appliedOrder"
           :orderDirection="appliedOrderDirection"
@@ -226,6 +229,9 @@ export default {
     },
     isNotificationVisible () {
       return this.$store.getters['user/isNotificationVisible'];
+    },
+    isHomePage () {
+      return this.$router.currentRoute.name === 'home';
     }
   },
   methods: {
@@ -300,16 +306,16 @@ export default {
   margin-bottom: 36px;
 }
 
-.full-width {
-  width: 100%;
-}
-
 .grid__head--sticky {
-  top: 64px;
-}
+  top: 128px;
 
-.grid__head--sticky.is-user-notification {
-  top: 64px + $notification-warning__height;
+  &.in-homePage {
+    top: 64px;
+  }
+
+  &.has-userNotification {
+    top: 128px + $notification-warning__height;
+  }
 }
 
 .pagination-element {